requirements.txt
accelerate
argh
dacite
demucs
diffusers==0.9.0
flask
flask_cors

### changed by oleverse
## must use numpy < 2.0 because of the conflict related to transformers package
# numpy
numpy<2.0

pillow>=9.1.0
plotly
pydub
pysoundfile
scipy
soundfile
sox
streamlit>=1.18.0
torch
torchaudio
torchvision
transformers
